requirements/base.txt
# Bleeding edge Django
django==1.6.2

# Configuration
django-configurations==0.8
django-secure==1.0
django-cache-url==0.8.0
dj-database-url==0.2.2

# Forms
django-braces==1.3.1
django-crispy-forms==1.4.0
django-floppyforms==1.1.1

# Models
South==0.8.4
django-model-utils==2.0.2

# images
Pillow==2.3.0

# For user registration, either via email or social
# Well-built with regular release cycles!
django-allauth==0.15.0

# For the persistance stores
psycopg2==2.5.2

# Unicode slugification
unicode-slugify==0.1.1
django-autoslug==1.7.2

# Useful things
django-avatar==2.0


# Your custom requirements go here
djangorestframework==2.3.13
